+++ src/DatePicker/LocaleUtils.js
@@ -34,30 +34,32 @@ const locales = {
   da
 };
 
-export function formatDate(date, dateFormat, locale) {
-  return format(date, dateFormat, {locale: locales[locale]});
-}
+const getLocale = locale =>
+  typeof locale === 'string' ? locales[locale] : locale;
+
+export const formatDate = (date, dateFormat, locale) =>
+  format(date, dateFormat, {locale: getLocale(locale)});
 
 export default locale => ({
   formatMonthTitle: date => format(date, 'MMMM YYYY', {
-    locale: locales[locale]
+    locale: getLocale(locale)
   }),
 
   formatWeekdayShort: index => format(setDay(new Date(), index), 'dd', {
-    locale: locales[locale]
+    locale: getLocale(locale)
   }),
 
   formatWeekdayLong: index => format(setDay(new Date(), index), 'dddd', {
-    locale: locales[locale]
+    locale: getLocale(locale)
   }),
 
   formatDay: date =>
     format(date, 'ddd ll', {
-      locale: locales[locale]
+      locale: getLocale(locale)
     }),
 
   getMonths: () =>
     [...Array(12).keys()].map(i =>
-      format(new Date(2018, i), 'MMMM', {locale: locales[locale]})
+      format(new Date(2018, i), 'MMMM', {locale: getLocale(locale)})
     )
 });